Part4, exercises 4.18-4.23: Token based authentication added and tests refactored to comply with code changes.

+++ osa4/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -9,8 +8,7 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
-
-  const user = await User.findById(body.userId);
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -26,8 +24,17 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  const user = request.user;
+  const isValidUser = user.id.toString() === blog.user.toString();
+  if (isValidUser) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response
+      .status(401)
+      .json({ error: "Invalid user. Cannot delete blog." });
+  }
 });
 
 blogsRouter.put("/:id", async (req, res) => {
